+++ viabill/translations/da.php
@@ -90,7 +90,7 @@ $_MODULE['<{viabill}prestashop>capturepaymenthandler_cf8fc70a60481b761aff4872351
 $_MODULE['<{viabill}prestashop>apirequest_365777a6e4be971cec32a5c47bcafa44'] = 'ViaBill-tjenesten er nede i øjeblikket. Vent og opdater siden, eller kontakt ViaBill-support på merchants@viabill.com ';
 $_MODULE['<{viabill}prestashop>orderstatusapiservice_4e0dec937cbd919782bfd55c7dc31642'] = 'En uventet fejl opstod for ordrer %s. Status code %s';
 $_MODULE['<{viabill}prestashop>paymentoptionsbuilder_54a23c9edf58e8e81c0f104e80e298e9'] = 'Betal med ViaBill';
-$_MODULE['<{viabill}prestashop>paymentoptionsbuilder_74abebcac5f468cb376eb110f042786d'] = 'Betal med ViaBill - Prøv før du køber';
+$_MODULE['<{viabill}prestashop>paymentoptionsbuilder_74abebcac5f468cb376eb110f042786d'] = 'Betal med ViaBill';
 $_MODULE['<{viabill}prestashop>paymentmanagementtemplate_d9978999a407502fb2de0f93ae38964d'] = 'På denne ordre er kun %s understøttet som valuta';
 $_MODULE['<{viabill}prestashop>installer_16a179ec546b3cbd0638c6444c61856a'] = 'ViaBill: Betaling afventer';
 $_MODULE['<{viabill}prestashop>installer_5aae3f26a4016c2c0b6abe11a9561054'] = 'ViaBill: Betaling godkendt';